refacto minioAdapter metadata
